의사 진료 후 동물의 정보 수정 기능 작업 완료

+++ src/main/java/com/myProject/treatment/domain/animal/dto/AnimalDTO.java
@@ -9,14 +9,14 @@ public class AnimalDTO {
     private Long id;
     @NotEmpty(message = "반려동물의 이름은 필수 입니다.")
     private String name;
-    private int height;
-    private int weight;
+    private double height;
+    private double weight;
     private String type;
     private Long memberId;
 
     public AnimalDTO() {}
 
-    public AnimalDTO(String name, int height, int weight, String type, Long memberId) {
+    public AnimalDTO(String name, double height, double weight, String type, Long memberId) {
         this.name = name;
         this.height = height;
         this.weight = weight;
@@ -24,7 +24,7 @@ public class AnimalDTO {
         this.memberId = memberId;
     }
 
-    public AnimalDTO(Long id, String name, int height, int weight, String type, Long memberId) {
+    public AnimalDTO(Long id, String name, double height, double weight, String type, Long memberId) {
         this.id = id;
         this.name = name;
         this.height = height;
